APP SERVER - autostop based price now it allow decimal values

+++ APP/BACKEND/src/ChargingSession/routes.js
@@ -169,7 +169,7 @@ router.post('/UpdateAutoStopSettings', async function(req, res) {
             updateFields.autostop_unit = parseFloat(req.body.updateUserUnitVal);
         }
         if (req.body.updateUserPriceVal !== undefined) {
-            updateFields.autostop_price = parseInt(req.body.updateUserPriceVal);
+            updateFields.autostop_price = parseFloat(req.body.updateUserPriceVal);
         }
         if (req.body.updateUserTime_isChecked !== undefined) {
             updateFields.autostop_time_is_checked = req.body.updateUserTime_isChecked;
